Changed - refactored WorkspaceView and WorkspaceState
